Project_BrainerHub/requirements.txt
djangorestframework==3.14.0
openpyxl==3.1.5
pandas==2.2.2
xlrd==2.0.1